+++ prp/models/phenotype.py
@@ -20,6 +20,7 @@ class PredictionSoftware(Enum):
     AMRFINDER = "amrfinder"
     RESFINDER = "resfinder"
     VIRFINDER = "virulencefinder"
+    SEROTYPEFINDER = "serotypefinder"
     MYKROBE = "mykrobe"
     TBPROFILER = "tbprofiler"
 
@@ -38,6 +39,7 @@ class ElementType(Enum):
     AMR = "AMR"
     STRESS = "STRESS"
     VIR = "VIRULENCE"
+    ANTIGEN = "ANTIGEN"
 
 
 class ElementStressSubtype(Enum):
@@ -50,9 +52,10 @@ class ElementStressSubtype(Enum):
 
 
 class ElementAmrSubtype(Enum):
-    """Categories of resistance and virulence genes."""
+    """Categories of resistance genes."""
 
     AMR = "AMR"
+    POINT = "POINT"
 
 
 class ElementVirulenceSubtype(Enum):
@@ -61,6 +64,12 @@ class ElementVirulenceSubtype(Enum):
     VIR = "VIRULENCE"
     ANTIGEN = "ANTIGEN"
     TOXIN = "TOXIN"
+
+
+class ElementSerotypeSubtype(Enum):
+    """Categories of serotype genes."""
+
+    ANTIGEN = "ANTIGEN"
 
 
 class PhenotypeInfo(RWModel):
@@ -95,7 +104,7 @@ class GeneBase(BaseModel):
         description="The predominant function fo the gene."
     )
     element_subtype: Union[
-        ElementStressSubtype, ElementAmrSubtype, ElementVirulenceSubtype
+        ElementStressSubtype, ElementAmrSubtype, ElementVirulenceSubtype, ElementSerotypeSubtype
     ] = Field(description="Further functional categorization of the genes.")
     # position
     ref_start_pos: Optional[int] = Field(
@@ -141,11 +150,15 @@ class ResistanceGene(GeneBase):
     phenotypes: List[PhenotypeInfo] = []
 
 
+class SerotypeGene(GeneBase):
+    """Container for serotype gene information"""
+
+
 class VirulenceGene(GeneBase, DatabaseReference):
     """Container for virulence gene information"""
 
     depth: Optional[float] = Field(
-        None, description="Ammount of sequence data supporting the gene."
+        None, description="Amount of sequence data supporting the gene."
     )
 
 
@@ -153,7 +166,7 @@ class ResfinderGene(ResistanceGene):
     """Container for Resfinder gene prediction information"""
 
     depth: Optional[float] = Field(
-        None, description="Ammount of sequence data supporting the gene."
+        None, description="Amount of sequence data supporting the gene."
     )